+++ packages/angular/angular-sdk/src/utils/getSpace.ts
@@ -3,19 +3,16 @@ import authenticate from './authenticate'
 import { Flatfile } from '@flatfile/api'
 
 type GetSpaceReturn = {
-  space: Flatfile.SpaceResponse
-  workbook: Flatfile.Workbook[]
+  space: Flatfile.Space
 }
 
 const getSpace = async (spaceProps: ISpace): Promise<GetSpaceReturn> => {
   const {
     space,
     apiUrl,
-    environmentId,
     spaceUrl = 'https://platform.flatfile.com/s/',
   } = spaceProps
   let spaceResponse
-  let workbookResponse
 
   try {
     if (!space?.id) {
@@ -25,16 +22,9 @@ const getSpace = async (spaceProps: ISpace): Promise<GetSpaceReturn> => {
       throw new Error('Missing required accessToken for Space')
     }
 
-    if (!environmentId) {
-      throw new Error('Missing required environment id')
-    }
-
     const limitedAccessApi = authenticate(space?.accessToken, apiUrl)
     try {
       spaceResponse = await limitedAccessApi.spaces.get(space?.id)
-      workbookResponse = await limitedAccessApi.workbooks.list({
-        spaceId: space?.id,
-      })
     } catch (error) {
       throw new Error(`Failed to get space: ${getErrorMessage(error)}`)
     }
@@ -48,7 +38,7 @@ const getSpace = async (spaceProps: ISpace): Promise<GetSpaceReturn> => {
       spaceResponse.data.guestLink = guestLink
     }
 
-    return { space: spaceResponse, workbook: workbookResponse.data }
+    return { space: spaceResponse.data }
   } catch (error) {
     const message = getErrorMessage(error)
     console.error(`Failed to initialize space: ${message}`)