+++ packages/embedded-utils/src/types/initSpace.ts
@@ -1,0 +1,5 @@
+import { ISpace } from '.'
+
+export type InitSpaceType = ISpace & {
+  isAutoConfig: boolean
+}